Migrate to vite + vitest + pnpm
We want to provide candidates a nice setup so they have a good image of
IMC, and can concentrate on business logic implementation. We don't want
candidates to have to waste time migrating the tooling.

CRA has a lack of maintenance and provides bad DX, as it's very slow for
today's standards. So let's use Vite as the main tool for project
development. Vite is one of the most popular and well-supported tools
nowadays with great DX. It has a fast dev-server and it's easy to
integrate other tooling with vite plugins.

As a test-runner, we go with vitest. It's a great addition to vite,
requiring less configuration that jest, and it supports ESM by default,
where jest is currently lacking or still experimental.

To be more aligned with our tech stack, let's use pnpm instead of yarn.