Provide a simple example of test with @testing-library
One of the requirements of the assignment is to write tests.
The react testing library is the industry standard tool used to test
components in isolation and in integration. We also use it at IMC.
Let's make it easier to write tests for the candidate by adding an
example that can be used as inspiration.